1. Quando un post o prodotto viene eliminato, viene eliminata anche la sua immagine dal file system

+++ controllers/PostController.js
@@ -1,5 +1,7 @@
 const PostModel = require('../models/postModel.js');
 const User      = require('../models/user.js');
+const fs        = require('fs');
+const path      = require('path');
 
 // creating a post
 
@@ -47,19 +49,36 @@ const updatePost = async (req, res) => {
 
 // delete a post
 const deletePost = async (req, res) => {
+
   const id = req.params.id;
-  //const { userId } = req.body;
 
   try {
+
     const post = await PostModel.findById(id);
-    //if (post.userId === userId) {
-      await post.deleteOne();
-      res.status(200).json("Post deleted.");
-    // } else {
-    //   res.status(403).json("Action forbidden");
-    // }
+
+	if (post === null) {
+		throw new Error(`The post with id ${id} doesn't exist`);
+	}
+
+	if (post.image) {
+
+		const imagePath = path.join(
+			global.rootDir, 
+			'public/images/post/', 
+			path.basename(post.image)
+		);
+
+		if (fs.existsSync(imagePath)) {
+			fs.unlinkSync(imagePath);
+		}
+	}
+
+    await post.deleteOne();
+
+    res.status(200).json("Post deleted.");
+
   } catch (error) {
-    res.status(400).json(error);
+    res.status(400).json(error.message);
   }
 };
